+++ ddtrace/contrib/aioredis/__init__.py
@@ -76,8 +76,11 @@ required_modules = ["aioredis"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.aiohttp.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        from ..internal.aioredis.patch import get_version
+        from ..internal.aioredis.patch import patch
+        from ..internal.aioredis.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
