+++ ddtrace/contrib/algoliasearch/__init__.py
@@ -29,8 +29,11 @@ required_modules = ["algoliasearch", "algoliasearch.version"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.aiohttp.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        from ..internal.algoliasearch.patch import get_version
+        from ..internal.algoliasearch.patch import patch
+        from ..internal.algoliasearch.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
