+++ ddtrace/contrib/anthropic/__init__.py
@@ -87,10 +87,11 @@ required_modules = ["anthropic"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from . import patch as _patch
+        # Required to allow users to import from `ddtrace.contrib.aiohttp.patch` directly
+        from . import patch as _  # noqa: F401, I001
 
-        patch = _patch.patch
-        unpatch = _patch.unpatch
-        get_version = _patch.get_version
+        from ..internal.anthropic.patch import patch
+        from ..internal.anthropic.patch import unpatch
+        from ..internal.anthropic.patch import get_version
 
         __all__ = ["patch", "unpatch", "get_version"]
