+++ ddtrace/contrib/asgi/__init__.py
@@ -62,8 +62,11 @@ required_modules = []
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .middleware import TraceMiddleware
-        from .middleware import get_version
-        from .middleware import span_from_scope
+        # Required to allow users to import from `ddtrace.contrib.aiohttp.patch` directly
+        from . import middleware as _  # noqa: F401, I001
+
+        from ..internal.asgi.middleware import TraceMiddleware
+        from ..internal.asgi.middleware import get_version
+        from ..internal.asgi.middleware import span_from_scope
 
         __all__ = ["TraceMiddleware", "span_from_scope", "get_version"]
