+++ ddtrace/contrib/asgi/utils.py
@@ -1,82 +1,15 @@
-"""
-Compatibility functions vendored from asgiref
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
+from ddtrace.vendor.debtcollector import deprecate
 
-Source: https://github.com/django/asgiref
-Version: 3.2.10
-License:
-
-Copyright (c) Django Software Foundation and individual contributors.
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification,
-are permitted provided that the following conditions are met:
-
-    1. Redistributions of source code must retain the above copyright notice,
-       this list of conditions and the following disclaimer.
-
-    2. Redistributions in binary form must reproduce the above copyright
-       notice, this list of conditions and the following disclaimer in the
-       documentation and/or other materials provided with the distribution.
-
-    3. Neither the name of Django nor the names of its contributors may be used
-       to endorse or promote products derived from this software without
-       specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
-ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
-ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""
-import asyncio
-import inspect
+from ..internal.asgi.utils import *  # noqa: F401,F403
 
 
-def is_double_callable(application):
-    """
-    Tests to see if an application is a legacy-style (double-callable) application.
-    """
-    # Look for a hint on the object first
-    if getattr(application, "_asgi_single_callable", False):
-        return False
-    if getattr(application, "_asgi_double_callable", False):
-        return True
-    # Uninstanted classes are double-callable
-    if inspect.isclass(application):
-        return True
-    # Instanted classes depend on their __call__
-    if hasattr(application, "__call__"):  # noqa: B004
-        # We only check to see if its __call__ is a coroutine function -
-        # if it's not, it still might be a coroutine function itself.
-        if asyncio.iscoroutinefunction(application.__call__):
-            return False
-    # Non-classes we just check directly
-    return not asyncio.iscoroutinefunction(application)
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
 
-
-def double_to_single_callable(application):
-    """
-    Transforms a double-callable ASGI application into a single-callable one.
-    """
-
-    async def new_application(scope, receive, send):
-        instance = application(scope)
-        return await instance(receive, send)
-
-    return new_application
-
-
-def guarantee_single_callable(application):
-    """
-    Takes either a single- or double-callable application and always returns it
-    in single-callable style. Use this to add backwards compatibility for ASGI
-    2.0 applications to your server/test harness/etc.
-    """
-    if is_double_callable(application):
-        application = double_to_single_callable(application)
-    return application
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
